+++ src/handlers/tsonBigint.ts
@@ -1,0 +1,7 @@
+import { TsonType } from "../types.js";
+
+export const tsonBigint: TsonType<bigint, string> = {
+	deserialize: (v) => BigInt(v),
+	primitive: "bigint",
+	serialize: (v) => v.toString(),
+};
