+++ src/handlers/tsonDate.ts
@@ -1,0 +1,7 @@
+import { TsonType } from "../types.js";
+
+export const tsonDate: TsonType<Date, string> = {
+	deserialize: (value) => new Date(value),
+	serialize: (value) => value.toJSON(),
+	test: (value) => value instanceof Date,
+};
