+++ src/handlers/tsonRegExp.ts
@@ -1,0 +1,11 @@
+import { TsonType } from "../types.js";
+
+export const tsonRegExp: TsonType<RegExp, string> = {
+	deserialize: (str) => {
+		const body = str.slice(1, str.lastIndexOf("/"));
+		const flags = str.slice(str.lastIndexOf("/") + 1);
+		return new RegExp(body, flags);
+	},
+	serialize: (value) => "" + value,
+	test: (value) => value instanceof RegExp,
+};
